wip(crud builder): refactoring
